Add support for managed bare metal platform.
This patch introduces support for the managed bare metal platform that
is currently under development under the metalkube github org.

More information about this platform type can be found in
documentation under https://github.com/metalkube/metalkube-docs/.

The actuator and other supporting components are still under active
devleopment, so we just point to an upstream metalkube image for the
actuator for now.  Having this in the machine-api-operator earlier
will make end-to-end testing a bit easier as the platform continues to
be developed.

Signed-off-by: Russell Bryant <rbryant@redhat.com>